Update solution for matrix-transpose

+++ matrix-transpose/matrix-transpose.py
@@ -4,5 +4,4 @@ def matrix_transpose(A):
     """
     Return the transpose of matrix A (swap rows and columns).
     """
-    # Write code here
-    return np.asarray(A, dtype=int).T
+    return np.asarray(A).T
